Added WithKeyPrefix extension for ISubscriber
