+++ StackExchange.Redis.Tests/TransactionWrapperTests.cs
@@ -11,7 +11,7 @@ namespace StackExchange.Redis.Tests
         private Mock<ITransaction> mock;
         private TransactionWrapper wrapper;
 
-        [TestFixtureSetUp]
+        [SetUp]
         public void Initialize()
         {
             mock = new Mock<ITransaction>();
