+++ StackExchange.Redis/StackExchange/Redis/KeyspaceIsolation/BatchWrapper.cs
@@ -2,7 +2,7 @@
 
 namespace StackExchange.Redis.StackExchange.Redis.KeyspaceIsolation
 {
-    internal sealed class BatchWrapper : WrapperBase<IBatch>, IBatch
+    internal sealed class BatchWrapper : DatabaseWrapperBase<IBatch>, IBatch
     {
         public BatchWrapper(IBatch inner, RedisKey prefix)
             : base(inner, prefix)
